✨ 구피 르 그란데 3페이즈 완료

+++ My_WinAPI/CEngine.cpp
@@ -125,7 +125,7 @@ HDC CEngine::GetSubDC()
 void CEngine::ChangeWindowSize(Vec2 _Resolution, bool _bMenu)
 {
 	// 입력된 해상도를 가져가기 위한 실제 윈도우의 크기를 계산
-	RECT rt = { 0, 0, _Resolution.x, _Resolution.y };
+	RECT rt = { 0, 0, (LONG)_Resolution.x, (LONG)_Resolution.y };
 	AdjustWindowRect(&rt, WS_OVERLAPPEDWINDOW, _bMenu);
 
 	// 윈도우 크기를 변경
